+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - widgets
     - fader: add `touchZone` property to allow restricting interaction to a part of the widget
+    - eq: add support for n-poles lowpass and highpass filters with the filter's `poles` property
+    - eq: adapt sampling frequency to the x-axis range
 
 ## 1.27.5